move some helper methods into a shared file
